place yaml_write_handler in distinct header file, style fixes
