+++ projects/libyaml/libyaml_reformatter_fuzzer.c
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include "yaml.h"
+#include "yaml_write_handler.h"
 #include <assert.h>
 #include <stdbool.h>
 #include <stdint.h>
@@ -23,28 +24,6 @@
 #ifdef NDEBUG
 #undef NDEBUG
 #endif
-
-typedef struct yaml_output_buffer {
-  unsigned char *buf;
-  size_t size;
-};
-
-static int yaml_write_handler(void *data, unsigned char *buffer, size_t size) {
-  struct yaml_output_buffer *out = (struct yaml_output_buffer *)data;
-  if (!out->size) {
-    out->buf = (unsigned char *)malloc(sizeof(unsigned char) * size);
-  } else {
-    out->buf = (unsigned char *)realloc(out->buf, out->size + size);
-  }
-
-  if (!out->buf) {
-    return 1;
-  }
-
-  memcpy(out->buf + out->size, buffer, size);
-  out->size += size;
-  return 0;
-}
 
 int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
   if (size < 2)
@@ -74,7 +53,7 @@ int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
   yaml_parser_set_input_string(&parser, data, size);
 
   /* Set the emitter parameters. */
-  struct yaml_output_buffer out = {/*buf=*/NULL, /*size=*/0};
+  yaml_output_buffer_t out = {/*buf=*/NULL, /*size=*/0};
   yaml_emitter_set_output(&emitter, yaml_write_handler, &out);
 
   yaml_emitter_set_canonical(&emitter, is_canonical);
@@ -102,9 +81,7 @@ int LLVMFuzzerTestOneInput(const uint8_t *data, size_t size) {
 
 error:
 
-  if (out.buf) {
-    free(out.buf);
-  }
+  free(out.buf);
 
   yaml_parser_delete(&parser);
   yaml_emitter_delete(&emitter);
